feat: Login Page redesign, Refactored code and CSS of other components
