+++ src/App.css
@@ -27,7 +27,7 @@
 
 .switch {
   display: flex;
-
+  margin-top: 7px;
 }
 
 /* App container */
@@ -58,7 +58,7 @@ nav div a {
 section {
   max-width: 500px;
   margin: 0 auto;
-  margin-top: 100px;
+  margin-top: 10%;
   box-shadow: 1px 2px 10px var(--color-4);
   padding: 10px 20px;
   border-radius: 5px;
@@ -68,8 +68,34 @@ section h3 {
   font-size: 20px;
   color: var(--color-4);
 }
+.logo {
+  width: 30px;
+  padding-right: 5px;
+}
+.logo_title {
+  display: flex;
+  font-weight: normal;
+  font-family:'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+}
 .login_div {
-  margin-left: 4%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+}
+.login_pic {
+  padding: 50px;
+}
+.login_buttons_div {
+  display: flex;
+  align-items: space-around;
+}
+.login_border {
+  border-bottom: 1px solid var(--color-6);
+  padding-top: 10px;
+  padding-bottom: 10px;
 }
 .logout_container{
   display: flex;
@@ -81,7 +107,6 @@ section h3 {
 }
 /* Register */
 section .form {
-  margin-top: 30px;
   padding: 0px 20px;
 }
 .input_container {
@@ -304,7 +329,7 @@ section .form {
 } 
 .users {
   overflow-y: auto;
-  height: 100%;
+  height: auto;
 }
 .users_container {
   overflow-y: auto;
@@ -318,11 +343,10 @@ section .form {
   height: auto;
 }
 .user_wrapper {
-  padding: 10px;
   cursor: pointer;
-  text-decoration: none;
 }
 .user_info {
+  text-decoration: none;
   display: flex;
   width: 100%;
   padding-left: 3%;
@@ -330,13 +354,13 @@ section .form {
 }
 .user_detail {
   display: flex;
-  width: 200px;
+  width: 100%;
   align-items: center;
 }
 .user_detail_div {
   margin-left: 10px;
-  margin-top: 10px;
-  max-width: 200px;
+  margin-top: 2px;
+  max-width: 100%;
   color: var(--color-2);
   white-space: nowrap;
   overflow: hidden;
@@ -344,7 +368,7 @@ section .form {
 }
 .avatar {
   width: 50px;
-  height: 50px;
+  height: 45px;
   border-radius: 50%;
 }
 .group_avatar_div {
@@ -390,12 +414,15 @@ section .form {
 .messages_user_back {
   font-size: 36px;
   text-decoration: none;
+  background:none; 
+  border:none;    
   display: hidden;
   visibility: hidden;
   color: var(--color-2);
   padding: 8px 16px;
 }
 .messages_user_back:hover {
+  background-color: #00000080;
   text-decoration: none;
   display: inline-block;
   padding: 8px 16px;
@@ -499,8 +526,6 @@ section .form {
 .truncate {
   font-size: 14px;
   white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
   flex: 1;
 }
 .unread {
@@ -551,6 +576,14 @@ section .form {
     visibility: visible;
     color: var(--color-2);
     padding: 8px 16px;
+  }
+  .login_popup_box {
+    width: 100vw;
+    height: 100vh;
+  }
+  .login_box {
+    width: 100vw;
+    height: 100vh;
   }
 }
 
@@ -633,6 +666,28 @@ section .form {
   justify-content: space-around;
   display: flex;
   flex-direction: column;
+}
+/* login-pop-up */
+.login_popup_box {
+  position: fixed;
+  display: flex;
+  align-items: center;
+  background: #00000080;;
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+}
+.login_box {
+  position: relative;
+  width: 70%;
+  margin: 0 auto;
+  height: 90%;
+  background: var(--color-8);
+  border-radius: 4px;
+  padding: 20px;
+  border: 1px solid #999;
+  overflow: auto;
 }
 
 /* width */
