+++ src/App.js
@@ -31,7 +31,7 @@ function App() {
               <Route exact path="/login" component={Login} />
               <PrivateRoute exact path="/profile" component={Profile} />
               <PrivateRoute exact path="/" component={Home} />
-              <Route exact path="/:convoId" component={Chat} />
+              <PrivateRoute exact path="/:convoId" component={Chat} />
             </Switch>
           </BrowserRouter>
         </AuthProvider>
