+++ src/components/Pop-ups/BalancePopUp.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 
-import { auth, db } from "../firebase";
+import { auth, db } from "../../firebase";
 import {
   Timestamp,
   addDoc,
