+++ src/components/Pop-ups/RegisterPopup.js
@@ -1,0 +1,98 @@
+import React, { useState } from "react";
+import { createUserWithEmailAndPassword } from "firebase/auth";
+import { auth, db } from "../../firebase";
+import { doc, Timestamp, setDoc } from "firebase/firestore";
+import { useHistory } from "react-router-dom";
+
+ 
+const RegisterPopup = props => {
+  const [data, setData] = useState({
+    name: "",
+    email: "",
+    password: "",
+    error: null,
+    loading: false,
+  });
+
+  const history = useHistory();
+
+  const { name, email, password, error, loading } = data;
+
+  const handleChange = (e) => {
+    setData({ ...data, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setData({ ...data, error: null, loading: true });
+    if (!name || !email || !password) {
+      setData({ ...data, error: "All fields are required" });
+    }
+    try {
+      const result = await createUserWithEmailAndPassword(
+        auth,
+        email,
+        password
+      );
+      await setDoc(doc(db, "users", result.user.uid), {
+        uid: result.user.uid,
+        name,
+        email,
+        createdAt: Timestamp.fromDate(new Date()),
+        isOnline: true,
+      });
+      setData({
+        name: "",
+        email: "",
+        password: "",
+        error: null,
+        loading: false,
+      });
+      history.replace("/");
+    } catch (err) {
+      setData({ ...data, error: err.message, loading: false });
+    }
+  };
+  return (
+    <div className="login_popup_box">
+      <div className="login_box">
+      <span className="close_icon" onClick={props.handleClose}>x</span>
+        <section>
+          <h3>Create An Account</h3>
+          <form className="form" onSubmit={handleSubmit}>
+            <div className="input_container">
+              <label htmlFor="name">Name</label>
+              <input type="text" name="name" value={name} onChange={handleChange} />
+            </div>
+            <div className="input_container">
+              <label htmlFor="email">Email</label>
+              <input
+                type="text"
+                name="email"
+                value={email}
+                onChange={handleChange}
+              />
+            </div>
+            <div className="input_container">
+              <label htmlFor="password">Password</label>
+              <input
+                type="password"
+                name="password"
+                value={password}
+                onChange={handleChange}
+              />
+            </div>
+            {error ? <p className="error">{error}</p> : null}
+            <div className="btn_container">
+              <button className="btn" disabled={loading}>
+                {loading ? "Creating ..." : "Register"}
+              </button>
+            </div>
+          </form>
+        </section>
+      </div>
+    </div>
+  );
+};
+ 
+export default RegisterPopup;
